Fixed WYSIWYG image upload not working without form URL provided
